refactor: cleanup shared modules
Got rid of remix/index.ts which mixed server and client code.
Dropped unused crypto utilities and jose dependency.
