+++ apps/builder/app/routes/dashboard.tsx
@@ -1,5 +1,5 @@
 import type { ServerRuntimeMetaFunction as MetaFunction } from "@remix-run/server-runtime";
-import { Root } from "~/shared/remix";
+import { Root } from "~/shared/remix/root";
 
 export const meta: MetaFunction = () => {
   return [{ title: "Webstudio Dashboard" }];
