--- apps/builder/app/shared/remix/index.ts
@@ -1,2 +1,0 @@
-export * from "./handle-request.server";
-export * from "./root";
